+++ lib/grape/validations/types/file.rb
@@ -18,9 +18,9 @@ module Grape
         def value_coerced?(value)
           # Rack::Request creates a Hash with filename,
           # content type and an IO object. Grape wraps that
-          # using hashie for convenience. Do a bit of basic
-          # duck-typing.
-          value.is_a?(Hashie::Mash) && value.key?(:tempfile)
+          # using ActiveSupport::HashWithIndifferentAccess for convenience.
+          # Do a bit of basic duck-typing.
+          value.is_a?(::Hash) && value.key?(:tempfile)
         end
       end
     end